requirements.txt
gradio
numpy
requests
groq
python-dotenv
beautifulsoup4


